frontend's structure of the properties revised
frontend's structure of the properties revised to support new mirrorjs
(0.1.2) implementations. Now the backend of a widget can ask
frontend-side properties. This is useful for example in the combobox to
get the number of elements or an element at a specified index +
implemented 2 methods in the combobox(.js) using this new feature:
"getItemsCount", "getItem"